range now is a Node field + some interface rearrangement
