+++ src/main/kotlin/astminer/common/model/Normalization.kt
@@ -1,0 +1,5 @@
+package astminer.common.model
+
+interface Normalization {
+    fun normalizeToken(token: String?): String
+}
