+++ src/main/kotlin/astminer/common/model/Token.kt
@@ -1,8 +1,9 @@
 package astminer.common.model
 
+import astminer.common.Code2VecNormalization
+
 class Token(
     val original: String?,
-    val range: TokenRange?,
     private val normalization: Normalization = Code2VecNormalization
 ) {
     val final: String
@@ -14,8 +15,3 @@ class Token(
 
     override fun toString(): String = final
 }
-
-typealias Line = Int
-typealias Column = Int
-
-data class TokenRange(val start: Pair<Line, Column>, val end: Pair<Line, Column>)
