+++ src/main/kotlin/astminer/parse/antlr/AntlrNode.kt
@@ -1,12 +1,15 @@
 package astminer.parse.antlr
 
 import astminer.common.model.Node
+import astminer.common.model.NodeRange
 
 class AntlrNode(
     override val typeLabel: String,
     override var parent: AntlrNode?,
     originalToken: String?
 ) : Node(originalToken) {
+
+    override val range: NodeRange? = null
 
     override val children: MutableList<AntlrNode> = mutableListOf()
