+++ src/main/kotlin/astminer/parse/treesitter/java/TreeSitterJavaFactory.kt
@@ -1,8 +1,8 @@
 package astminer.parse.treesitter.java
 
+import astminer.common.SimpleNode
 import astminer.common.model.ParsingResult
 import astminer.common.model.ParsingResultFactory
-import astminer.common.model.SimpleNode
 import astminer.common.model.TreeFunctionSplitter
 import java.io.File
