+++ src/main/kotlin/astminer/parse/treesitter/java/TreeSitterJavaFunctionInfo.kt
@@ -1,7 +1,8 @@
 package astminer.parse.treesitter.java
 
+import astminer.common.Code2VecNormalization
+import astminer.common.SimpleNode
 import astminer.common.model.*
-import astminer.common.model.Code2VecNormalization
 import astminer.parse.antlr.getTokensFromSubtree
 import astminer.parse.findEnclosingElementBy
 import mu.KotlinLogging
