+++ src/test/kotlin/astminer/common/Code2VecNormalizationTest.kt
@@ -1,6 +1,5 @@
 package astminer.common
 
-import astminer.common.model.Code2VecNormalization
 import org.junit.Assert
 import org.junit.Test
